+++ src/util/string_tools.cpp
@@ -60,43 +60,6 @@ void formatHeaderCasing(std::string& header) {
     }
 }
 
-int compressText(std::string& buffer, const int method) {
-    if (method == COMPRESS_BROTLI) {
-        buffer = brotli::compress(buffer);
-    } else {
-        // Handle deflate compression
-        const size_t sourceLen = buffer.size();
-        z_stream zs;
-        zs.zalloc = Z_NULL;
-        zs.zfree = Z_NULL;
-        zs.opaque = Z_NULL;
-
-        const int windowBits = method == COMPRESS_GZIP ? (15 | 16) : 15; 
-        if (deflateInit2(&zs, Z_DEFAULT_COMPRESSION, Z_DEFLATED, windowBits, 8, Z_DEFAULT_STRATEGY) != Z_OK)
-            return IO_FAILURE;
-
-        // Determine output size
-        const size_t outputLen = deflateBound(&zs, sourceLen);
-        char* pBuffer = new char[outputLen];
-
-        zs.avail_in = (uInt)sourceLen;
-        zs.avail_out = (uInt)outputLen;
-        zs.next_in = (Bytef*)buffer.c_str();
-        zs.next_out = (Bytef*)pBuffer;
-
-        // Deflate
-        deflate(&zs, Z_FINISH);
-        deflateEnd(&zs);
-
-        // Copy buffer
-        buffer.clear();
-        for (size_t i = 0; i < zs.total_out; ++i)
-            buffer.push_back(pBuffer[i]);
-        delete[] pBuffer;
-    }
-    return IO_SUCCESS;
-}
-
 bool isMostlyAscii(const std::string& data, double thresh) {
     int asciiCount = 0;